Updated to R-devel rev 72600.

+++ src/unix/sys-std.c
@@ -1,7 +1,7 @@
 /*
  *  R : A Computer Language for Statistical Data Analysis
  *  Copyright (C) 1995, 1996  Robert Gentleman and Ross Ihaka
- *  Copyright (C) 1997--2016  The R Core Team
+ *  Copyright (C) 1997--2017  The R Core Team
  *  Copyright (C) 2008-2014  Andrew R. Runnalls.
  *  Copyright (C) 2014 and onwards the Rho Project Authors.
  *
@@ -962,7 +962,7 @@ Rstd_ReadConsole(const char *prompt, unsigned char *buf, int len,
 /* according to system.txt, should be terminated in \n, so check this
    at eof and error */
 	if ((err || feof(ifp ? ifp : stdin))
-	    && (ll == 0 || buf[ll - 1] != '\n') && ll < len) {
+	    && (ll == 0 || buf[ll - 1] != '\n') && ll < (size_t)len) {
 	    buf[ll++] = '\n'; buf[ll] = '\0';
 	}
 	if (!R_Slave) {
